Refactored server actions to be in separate files: * Created actions/data-intake.ts * Created actions/journal.ts * Moved /utils/supabase to be /supabase, deleted /utils altogether * Refactored all imports as necessary * Re-ran tests and verified that they still work * Manually tests many of the features locally
