+++ src/app/home/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from 'next/navigation';
-import { createClient } from '@/utils/supabase/server';
+import { createClient } from '@/supabase/server';
 import Footer from '@/components/footer';
 import { Header } from '@/components/header';
 import Dashboard from '@/components/dashboard';
