--- src/lib/types.ts
@@ -1,1 +1,0 @@
-// TODO: Add supabase table types!
