+++ src/supabase/scripts/handle_new_user.sql
