+++ src/supabase/scripts/sync_auth_user_data.sql
