+++ src/supabase/scripts/tables.sql
